Implement combine reducers functionality
