+++ src/actions/index.js
@@ -1,11 +1,14 @@
+const CREATE_BOOK = 'CREATE_BOOK';
+const REMOVE_BOOK = 'REMOVE_BOOK';
+
 const createBook = (book) => ({
-    type: 'CREATE_BOOK',
-    book,
-  });
-  
-const removeBook = (book) => ({
-    type: 'REMOVE_BOOK',
-    book,
+  type: CREATE_BOOK,
+  book,
 });
 
-export {createBook, removeBook};
+const removeBook = (book) => ({
+  type: REMOVE_BOOK,
+  book,
+});
+
+export { createBook, removeBook };
