+++ src/containers/BooksList.js
@@ -5,6 +5,8 @@ import Book from '../components/Book';
 const BooksList = (props) => {
   const { books } = props;
 
+  console.log(props);
+
   return (
     <div className="BooksList">
       <table>
@@ -39,7 +41,7 @@ BooksList.propTypes = {
   })),
 };
 
-const mapStateToProps = (state) => ({ books: state });
+const mapStateToProps = (state) => (state);
 
 const mapDispatchToProps = (dispatch) => ({
   submitNewBook: () => {
