+++ src/reducers/books.js
@@ -1,0 +1,31 @@
+const initialState = [
+  {
+    id: 1,
+    title: 'Harry Porter series',
+    category: 'Fantasy',
+  },
+  {
+    id: 2,
+    title: 'The New Man',
+    category: 'Fiction',
+  },
+];
+
+const booksReducer = (state = initialState, action) => {
+  let output;
+  switch (action.type) {
+    case 'CREATE_BOOK':
+      output = 'CREATED';
+      break;
+    case 'REMOVE_BOOK':
+      output = 'CREATED';
+      break;
+    default:
+      output = state;
+  }
+
+  console.log(output);
+  return output;
+};
+
+export default booksReducer;
