+++ src/reducers/index.js
@@ -1,0 +1,8 @@
+import { combineReducers } from 'redux';
+import booksReducer from './books';
+
+const rootReducer = combineReducers({
+  books: booksReducer,
+});
+
+export default rootReducer;
